requirements.txt
numpy>=1.13.3         #BSD 3-Clause
requests>=2.18.4      # Apache 2.0
#japanize_kivy==0.1.1
#Kivy==1.10.1
matplotlib>=3.1.3     #Python Software Foundation License
PyYAML>=5.4.1         #MIT
google-api-python-client>=2.71.0